+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 # 1.0.0
 ```
+    * When call Checkokut SDK, pass the billing address and into userData block.
+    * Removed nuveiLoadCheckout() method from frontend JS.
+    * Removed nuveiGetSessionToken() method from frontend JS.
     * Removed the Cron job for a new plugin version.
     * In this plugin we only replace the Web SDK with the Checkout SDK.
 ```